requirements.txt
imageai==2.0.2
pyFirmata==1.0.3
numpy==1.15.4
tornado==5.1
opencv_contrib_python==3.4.2.17
Pillow==5.3.0
pyserial==3.4
scikit_learn==0.20.0
matplotlib==2.2.3
mahotas==1.4.4
imutils==0.5.1
h5py==2.8.0
scikit_image==0.14.1
